+++ rustconn/src/embedded_rdp.rs
@@ -3135,6 +3135,42 @@ impl EmbeddedRdpWidget {
                                     .send(RdpClientCommand::ClipboardData { format_id, data });
                             }
                         }
+                        RdpClientEvent::ClipboardFileList(files) => {
+                            // File list available on server clipboard
+                            tracing::info!("[Clipboard] File list received: {} files", files.len());
+                            for file in &files {
+                                tracing::debug!(
+                                    "  - {} ({} bytes, dir={})",
+                                    file.name,
+                                    file.size,
+                                    file.is_directory()
+                                );
+                            }
+                            // TODO: Store file list for UI display and download
+                        }
+                        RdpClientEvent::ClipboardFileContents {
+                            stream_id,
+                            data,
+                            is_last,
+                        } => {
+                            // File contents received from server
+                            tracing::debug!(
+                                "[Clipboard] File contents: stream_id={}, {} bytes, last={}",
+                                stream_id,
+                                data.len(),
+                                is_last
+                            );
+                            // TODO: Write data to local file
+                        }
+                        RdpClientEvent::ClipboardFileSize { stream_id, size } => {
+                            // File size information received
+                            tracing::debug!(
+                                "[Clipboard] File size: stream_id={}, size={}",
+                                stream_id,
+                                size
+                            );
+                            // TODO: Use for progress indication
+                        }
                     }
                 }
             }